requirements.txt
lxml==5.3.0
pillow==11.0.0
